Move module file into src/groovis
